Browser extension code cleanup
- Refactor content script code to a more modular form
